add argparse and put funtion in utils.py
